Move deposit processing into BlockProcessor
