+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/util/CommitteeUtil.java
@@ -42,6 +42,7 @@ import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateCache;
 
+@Deprecated
 public class CommitteeUtil {
 
   /**
