+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/block/AbstractBlockProcessor.java
@@ -15,15 +15,21 @@ package tech.pegasys.teku.spec.logic.common.block;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static java.lang.Math.toIntExact;
+import static tech.pegasys.teku.spec.config.SpecConfig.FAR_FUTURE_EPOCH;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.OptionalInt;
+import java.util.function.Function;
+import java.util.stream.IntStream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.crypto.Hash;
+import tech.pegasys.teku.bls.BLS;
 import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.bls.BLSSignatureVerifier;
 import tech.pegasys.teku.bls.BLSSignatureVerifier.InvalidSignatureException;
@@ -38,6 +44,9 @@ import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBody;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
 import tech.pegasys.teku.spec.datastructures.operations.AttesterSlashing;
 import tech.pegasys.teku.spec.datastructures.operations.Deposit;
+import tech.pegasys.teku.spec.datastructures.operations.DepositData;
+import tech.pegasys.teku.spec.datastructures.operations.DepositMessage;
+import tech.pegasys.teku.spec.datastructures.operations.DepositWithIndex;
 import tech.pegasys.teku.spec.datastructures.operations.ProposerSlashing;
 import tech.pegasys.teku.spec.datastructures.operations.SignedVoluntaryExit;
 import tech.pegasys.teku.spec.datastructures.state.Validator;
@@ -46,6 +55,7 @@ import tech.pegasys.teku.spec.datastructures.state.beaconstate.MutableBeaconStat
 import tech.pegasys.teku.spec.datastructures.util.AttestationProcessingResult;
 import tech.pegasys.teku.spec.logic.common.helpers.BeaconStateAccessors;
 import tech.pegasys.teku.spec.logic.common.helpers.MiscHelpers;
+import tech.pegasys.teku.spec.logic.common.helpers.Predicates;
 import tech.pegasys.teku.spec.logic.common.operations.signatures.ProposerSlashingSignatureVerifier;
 import tech.pegasys.teku.spec.logic.common.operations.signatures.VoluntaryExitSignatureVerifier;
 import tech.pegasys.teku.spec.logic.common.operations.validation.AttesterSlashingStateTransitionValidator;
@@ -59,22 +69,33 @@ import tech.pegasys.teku.spec.logic.common.util.ValidatorsUtil;
 import tech.pegasys.teku.ssz.SszList;
 
 public abstract class AbstractBlockProcessor implements BlockProcessor {
+  /**
+   * For debug/test purposes only enables/disables {@link DepositData} BLS signature verification
+   * Setting to <code>false</code> significantly speeds up state initialization
+   */
+  public static boolean BLS_VERIFY_DEPOSIT = true;
+
   private static final Logger LOG = LogManager.getLogger();
+
   protected final SpecConfig specConfig;
-  protected final BeaconStateUtil beaconStateUtil;
-  protected final AttestationUtil attestationUtil;
-  protected final ValidatorsUtil validatorsUtil;
+  protected final Predicates predicates;
   protected final MiscHelpers miscHelpers;
   protected final BeaconStateAccessors beaconStateAccessors;
 
+  protected final BeaconStateUtil beaconStateUtil;
+  protected final AttestationUtil attestationUtil;
+  protected final ValidatorsUtil validatorsUtil;
+
   protected AbstractBlockProcessor(
       final SpecConfig specConfig,
+      final Predicates predicates,
+      final MiscHelpers miscHelpers,
+      final BeaconStateAccessors beaconStateAccessors,
       final BeaconStateUtil beaconStateUtil,
       final AttestationUtil attestationUtil,
-      final ValidatorsUtil validatorsUtil,
-      final MiscHelpers miscHelpers,
-      final BeaconStateAccessors beaconStateAccessors) {
+      final ValidatorsUtil validatorsUtil) {
     this.specConfig = specConfig;
+    this.predicates = predicates;
     this.beaconStateUtil = beaconStateUtil;
     this.attestationUtil = attestationUtil;
     this.validatorsUtil = validatorsUtil;
@@ -411,12 +432,109 @@ public abstract class AbstractBlockProcessor implements BlockProcessor {
       throws BlockProcessingException {
     try {
       for (Deposit deposit : deposits) {
-        beaconStateUtil.processDeposit(state, deposit);
+        processDeposit(state, deposit);
       }
     } catch (IllegalArgumentException e) {
       LOG.warn(e.getMessage());
       throw new BlockProcessingException(e);
     }
+  }
+
+  public void processDeposit(MutableBeaconState state, Deposit deposit) {
+    checkArgument(
+        predicates.isValidMerkleBranch(
+            deposit.getData().hashTreeRoot(),
+            deposit.getProof(),
+            specConfig.getDepositContractTreeDepth() + 1, // Add 1 for the List length mix-in
+            toIntExact(state.getEth1_deposit_index().longValue()),
+            state.getEth1_data().getDeposit_root()),
+        "process_deposit: Verify the Merkle branch");
+
+    processDepositWithoutCheckingMerkleProof(state, deposit, null);
+  }
+
+  @Override
+  public void processDepositWithoutCheckingMerkleProof(
+      final MutableBeaconState state,
+      final Deposit deposit,
+      final Map<BLSPublicKey, Integer> pubKeyToIndexMap) {
+    state.setEth1_deposit_index(state.getEth1_deposit_index().plus(UInt64.ONE));
+
+    final BLSPublicKey pubkey = deposit.getData().getPubkey();
+    final UInt64 amount = deposit.getData().getAmount();
+
+    OptionalInt existingIndex;
+    if (pubKeyToIndexMap != null) {
+      final Integer cachedIndex =
+          pubKeyToIndexMap.putIfAbsent(pubkey, state.getValidators().size());
+      existingIndex = cachedIndex == null ? OptionalInt.empty() : OptionalInt.of(cachedIndex);
+    } else {
+      SszList<Validator> validators = state.getValidators();
+
+      Function<Integer, BLSPublicKey> validatorPubkey =
+          index ->
+              beaconStateAccessors.getValidatorPubKey(state, UInt64.valueOf(index)).orElse(null);
+
+      existingIndex =
+          IntStream.range(0, validators.size())
+              .filter(index -> pubkey.equals(validatorPubkey.apply(index)))
+              .findFirst();
+    }
+
+    if (existingIndex.isEmpty()) {
+
+      // Verify the deposit signature (proof of possession) which is not checked by the deposit
+      // contract
+      if (BLS_VERIFY_DEPOSIT) {
+        final DepositMessage deposit_message =
+            new DepositMessage(pubkey, deposit.getData().getWithdrawal_credentials(), amount);
+        final Bytes32 domain = beaconStateUtil.computeDomain(specConfig.getDomainDeposit());
+        final Bytes signing_root = beaconStateUtil.computeSigningRoot(deposit_message, domain);
+        boolean proof_is_valid =
+            !BLS_VERIFY_DEPOSIT
+                || BLS.verify(pubkey, signing_root, deposit.getData().getSignature());
+        if (!proof_is_valid) {
+          if (deposit instanceof DepositWithIndex) {
+            LOG.debug(
+                "Skipping invalid deposit with index {} and pubkey {}",
+                ((DepositWithIndex) deposit).getIndex(),
+                pubkey);
+          } else {
+            LOG.debug("Skipping invalid deposit with pubkey {}", pubkey);
+          }
+          if (pubKeyToIndexMap != null) {
+            // The validator won't be created so the calculated index won't be correct
+            pubKeyToIndexMap.remove(pubkey);
+          }
+          return;
+        }
+      }
+
+      if (pubKeyToIndexMap == null) {
+        LOG.debug("Adding new validator to state: {}", state.getValidators().size());
+      }
+      state.getValidators().append(getValidatorFromDeposit(deposit));
+      state.getBalances().appendElement(amount);
+    } else {
+      validatorsUtil.increaseBalance(state, existingIndex.getAsInt(), amount);
+    }
+  }
+
+  private Validator getValidatorFromDeposit(Deposit deposit) {
+    final UInt64 amount = deposit.getData().getAmount();
+    final UInt64 effectiveBalance =
+        amount
+            .minus(amount.mod(specConfig.getEffectiveBalanceIncrement()))
+            .min(specConfig.getMaxEffectiveBalance());
+    return new Validator(
+        deposit.getData().getPubkey(),
+        deposit.getData().getWithdrawal_credentials(),
+        effectiveBalance,
+        false,
+        FAR_FUTURE_EPOCH,
+        FAR_FUTURE_EPOCH,
+        FAR_FUTURE_EPOCH,
+        FAR_FUTURE_EPOCH);
   }
 
   /**
