+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/helpers/Predicates.java
@@ -13,8 +13,12 @@
 
 package tech.pegasys.teku.spec.logic.common.helpers;
 
+import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.bytes.Bytes32;
+import org.apache.tuweni.crypto.Hash;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.datastructures.state.Validator;
+import tech.pegasys.teku.ssz.collections.SszBytes32Vector;
 
 public class Predicates {
 
@@ -30,5 +34,18 @@ public class Predicates {
   public boolean isActiveValidator(Validator validator, UInt64 epoch) {
     return validator.getActivation_epoch().compareTo(epoch) <= 0
         && epoch.compareTo(validator.getExit_epoch()) < 0;
+  }
+
+  public boolean isValidMerkleBranch(
+      Bytes32 leaf, SszBytes32Vector branch, int depth, int index, Bytes32 root) {
+    Bytes32 value = leaf;
+    for (int i = 0; i < depth; i++) {
+      if (Math.floor(index / Math.pow(2, i)) % 2 == 1) {
+        value = Hash.sha2_256(Bytes.concatenate(branch.getElement(i), value));
+      } else {
+        value = Hash.sha2_256(Bytes.concatenate(value, branch.getElement(i)));
+      }
+    }
+    return value.equals(root);
   }
 }
