+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/phase0/SpecLogicPhase0.java
@@ -98,11 +98,12 @@ public class SpecLogicPhase0 extends AbstractSpecLogic {
     final BlockProcessorPhase0 blockProcessor =
         new BlockProcessorPhase0(
             config,
+            predicates,
+            miscHelpers,
+            beaconStateAccessors,
             beaconStateUtil,
             attestationUtil,
-            validatorsUtil,
-            beaconStateAccessors,
-            miscHelpers);
+            validatorsUtil);
     final StateTransition stateTransition =
         StateTransition.create(
             config, blockProcessor, epochProcessor, beaconStateUtil, beaconStateAccessors);
