+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/versions/phase0/util/BlockProcessorPhase0.java
@@ -29,6 +29,7 @@ import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.phase0.M
 import tech.pegasys.teku.spec.logic.common.block.AbstractBlockProcessor;
 import tech.pegasys.teku.spec.logic.common.helpers.BeaconStateAccessors;
 import tech.pegasys.teku.spec.logic.common.helpers.MiscHelpers;
+import tech.pegasys.teku.spec.logic.common.helpers.Predicates;
 import tech.pegasys.teku.spec.logic.common.operations.validation.AttestationDataStateTransitionValidator;
 import tech.pegasys.teku.spec.logic.common.operations.validation.OperationInvalidReason;
 import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.BlockProcessingException;
@@ -42,18 +43,20 @@ public final class BlockProcessorPhase0 extends AbstractBlockProcessor {
 
   public BlockProcessorPhase0(
       final SpecConfig specConfig,
+      final Predicates predicates,
+      final MiscHelpers miscHelpers,
+      final BeaconStateAccessors beaconStateAccessors,
       final BeaconStateUtil beaconStateUtil,
       final AttestationUtil attestationUtil,
-      final ValidatorsUtil validatorsUtil,
-      final BeaconStateAccessors beaconStateAccessors,
-      final MiscHelpers miscHelpers) {
+      final ValidatorsUtil validatorsUtil) {
     super(
         specConfig,
+        predicates,
+        miscHelpers,
+        beaconStateAccessors,
         beaconStateUtil,
         attestationUtil,
-        validatorsUtil,
-        miscHelpers,
-        beaconStateAccessors);
+        validatorsUtil);
   }
 
   @Override
